Move the scale lock button from QgsStatusBarScaleWidget to QgsStatusBarMagnifierWidget, so it's hidable together with the magnifier in the Customization
