+++ src/app/qgsstatusbarmagnifierwidget.cpp
@@ -17,6 +17,7 @@
 #include <QFont>
 #include <QHBoxLayout>
 #include <QLabel>
+#include <QToolButton>
 
 #include "qgssettings.h"
 #include "qgsapplication.h"
@@ -55,8 +56,18 @@ QgsStatusBarMagnifierWidget::QgsStatusBarMagnifierWidget( QWidget *parent )
 
   connect( mSpinBox, static_cast < void ( QgsDoubleSpinBox::* )( double ) > ( &QgsDoubleSpinBox::valueChanged ), this, &QgsStatusBarMagnifierWidget::setMagnification );
 
+  mLockButton = new QToolButton();
+  mLockButton->setIcon( QIcon( QgsApplication::getThemeIcon( "/lockedGray.svg" ) ) );
+  mLockButton->setToolTip( tr( "Lock the scale to use magnifier to zoom in or out." ) );
+  mLockButton->setCheckable( true );
+  mLockButton->setChecked( false );
+  mLockButton->setAutoRaise( true );
+
+  connect( mLockButton, &QAbstractButton::toggled, this, &QgsStatusBarMagnifierWidget::scaleLockChanged );
+
   // layout
   mLayout = new QHBoxLayout( this );
+  mLayout->addWidget( mLockButton );
   mLayout->addWidget( mLabel );
   mLayout->addWidget( mSpinBox );
   mLayout->setContentsMargins( 0, 0, 0, 0 );
