+++ src/app/qgsstatusbarscalewidget.h
@@ -20,7 +20,6 @@
 class QFont;
 class QHBoxLayout;
 class QLabel;
-class QToolButton;
 class QValidator;
 
 class QgsMapCanvas;
@@ -47,6 +46,13 @@ class APP_EXPORT QgsStatusBarScaleWidget : public QWidget
     void setScale( double scale );
 
     /**
+     * Lock the scale widget.
+     * \param state the lock state
+     * \since QGIS 3.4
+     */
+    void setLocked( bool state );
+
+    /**
      * \brief isLocked check if the scale should be locked to use magnifier instead of scale to zoom in/out
      * \returns True if the scale shall be locked
      */
@@ -64,13 +70,9 @@ class APP_EXPORT QgsStatusBarScaleWidget : public QWidget
   private slots:
     void userScale() const;
 
-  signals:
-    void scaleLockChanged( bool );
-
   private:
     QgsMapCanvas *mMapCanvas = nullptr;
     QHBoxLayout *mLayout = nullptr;
-    QToolButton *mLockButton = nullptr;
 
     //! Widget that will live on the statusbar to display "scale 1:"
     QLabel *mLabel = nullptr;
